Expose MetadataCache through testkit framework, clean up some "raft" references
