+++ checkstyle/import-control-core.xml
@@ -95,6 +95,7 @@
     <subpackage name="junit">
       <allow pkg="kafka.test"/>
       <allow pkg="kafka.testkit"/>
+      <allow pkg="org.apache.kafka.image"/>
       <allow pkg="org.apache.kafka.clients"/>
       <allow pkg="org.apache.kafka.metadata" />
     </subpackage>
